+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Rework ProfileLinker Component (UIDATIMP-358)
 * Convert Match Profiles form Existing Record Field name to JSONPath format (UIDATIMP-375)
 * Define API Contract for MappingProfile field mapping definitions (UIDATIMP-377)
+* Augment RecordTypeSelector component with Incoming Record Type selection dropdown (UIDATIMP-386)
 
 ### Bugs fixed:
 * Profile Associator lists are empty when the user reloads the page with Profile Edit Form open (UIDATIMP-338)